+++ src/EntityGraphQL.AspNet/Extensions/EntityGraphQLEndpointRouteExtensions.cs
@@ -1,13 +1,11 @@
 using System.Linq;
 using System.Text;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 using EntityGraphQL.Schema;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-using Newtonsoft.Json.Serialization;
 
 namespace EntityGraphQL.AspNet.Extensions
 {
@@ -23,22 +21,22 @@ namespace EntityGraphQL.AspNet.Extensions
                 await context.Request.Body.ReadAsync(buffer, 0, buffer.Length);
                 var json = Encoding.UTF8.GetString(buffer);
 
-                var query = JsonConvert.DeserializeObject<QueryRequest>(json);
+                var jsonOptions = new JsonSerializerOptions
+                {
+                    IncludeFields = true,
+                    PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+                };
+                jsonOptions.Converters.Add(new JsonStringEnumConverter());
+                var query = JsonSerializer.Deserialize<QueryRequest>(json, jsonOptions);
                 var schema = context.RequestServices.GetService<SchemaProvider<TQueryType>>();
                 var schemaContext = context.RequestServices.GetService<TQueryType>();
                 var data = await schema.ExecuteQueryAsync(query, schemaContext, context.RequestServices, context.User.Identities.FirstOrDefault(), options);
-                var serializerSettings = new JsonSerializerSettings
-                {
-                    ContractResolver = new CamelCasePropertyNamesContractResolver()
-                };
-                serializerSettings.Converters.Add(new StringEnumConverter());
                 context.Response.ContentType = "application/json; charset=utf-8";
-                var jsonResult = JsonConvert.SerializeObject(data, serializerSettings);
+                var jsonResult = JsonSerializer.Serialize(data, jsonOptions);
                 var jsonBytes = Encoding.UTF8.GetBytes(jsonResult);
                 context.Response.ContentLength = jsonBytes.Length;
                 await context.Response.Body.WriteAsync(jsonBytes);
             });
-
 
             return builder;
         }