+++ src/demo/Mutations/DemoMutations.cs
@@ -125,7 +125,7 @@ namespace demo.Mutations
         public string Name { get; set; }
         public double Rating { get; set; }
         public DateTime Released;
-        public List<Detail> Details { get; set; }
+        public Detail Details { get; set; }
     }
 
     public class Detail